+++ vaspauto/__init__.py
@@ -5,3 +5,5 @@ Usage:
     python3 -m vaspauto submit -c config.toml -N 2 -n 112
     python3 -m vaspauto scheduler -c config.toml -n 112
 """
+
+__version__ = '5.3'